REF: Minimize state in read_csv (#59194)
* Move some set parse_date_cols to pythong parser

* Clean up do_date_conversions

* Move can cast to python parser

* Move can cast to python parser

* Revert "Move can cast to python parser"

This reverts commit 99ca747bbc4c82934245384049a209459432221b.

* Typing issues
